+++ Umbraco.AI/src/Umbraco.AI.Web.StaticAssets/Client/src/knowledge-set/constants.ts
@@ -1,0 +1,8 @@
+export { UAI_KNOWLEDGE_SET_ENTITY_TYPE, UAI_KNOWLEDGE_SET_ROOT_ENTITY_TYPE } from "./entity.js";
+export type { UaiKnowledgeSetEntityType, UaiKnowledgeSetRootEntityType } from "./entity.js";
+
+export * from "./repository/constants.js";
+export * from "./workspace/constants.js";
+export * from "./collection/constants.js";
+
+export const UAI_KNOWLEDGE_SET_ICON = "icon-books";